fix: change size() return type to size_t

+++ Team41/Code41/src/spa/src/PKB/ManyToMany.h
@@ -90,14 +90,14 @@ public:
      *
      * @return number of keys
      */
-    int keySize();
+    size_t keySize();
 
     /**
      * Gets the number of values stored in table
      *
      * @return number of vals
      */
-    int valSize();
+    size_t valSize();
 };
 
 #include "ManyToMany.tpp"
